update 09-07-2121
updated images to be lower pull on server, made images smaller and ran through optimizer.

added two new skills yootheme and photoshop and moved them to the top.

tweeked the design and fixed a couple aspects of the hover to make it look cleaner and fixed some margin and padding issues.